refactor: Restructure bento-get package for improved organization
Implementation reason: The restructuring of the bento-get package improves code organization, separates concerns, and enhances maintainability by breaking down large files into smaller, more focused components.

- Added new files:
  - AuthGuard.tsx
  - handle-global-keypress.ts
  - SearchBar.tsx
  - InstallConfirmScreen.tsx
  - SearchScreen.tsx
- Moved existing components to separate files
- Updated imports and exports accordingly
- Removed unused setInstallingComponent from AppContext
- Simplified main App component
- Created new type definition files for added components
