+++ code/packages/bento-get/src/screens/SearchScreen.tsx
@@ -1,0 +1,10 @@
+import { UsageBanner, ResultsContainer } from '../cli-components/index.js'
+import { SearchBar } from '../cli-components/SearchBar.js'
+
+export const SearchScreen = () => (
+  <>
+    <UsageBanner />
+    <SearchBar />
+    <ResultsContainer />
+  </>
+)
