+++ code/packages/bento-get/types/app/AuthGuard.d.ts
@@ -1,0 +1,5 @@
+import React from 'react';
+export declare const AuthGuard: ({ children }: {
+    children: React.ReactNode;
+}) => import("react/jsx-runtime").JSX.Element | undefined;
+//# sourceMappingURL=AuthGuard.d.ts.map
