+++ code/packages/bento-get/types/app/handle-global-keypress.d.ts
@@ -1,0 +1,5 @@
+import type { AppContextType } from '../data/AppContext.js';
+export declare const handleGlobalKeyPress: (key: string, modifier: any, appContext: AppContextType, navigate: (path: string) => void, location: {
+    pathname: string;
+}) => void;
+//# sourceMappingURL=handle-global-keypress.d.ts.map
