+++ code/packages/bento-get/types/cli-components/SearchBar.d.ts
@@ -1,0 +1,2 @@
+export declare const SearchBar: () => import("react/jsx-runtime").JSX.Element;
+//# sourceMappingURL=SearchBar.d.ts.map
