+++ code/packages/bento-get/types/components/index.d.ts
@@ -1,0 +1,21 @@
+export declare const ResultsContainer: () => import("react/jsx-runtime").JSX.Element;
+export declare const Footer: () => import("react/jsx-runtime").JSX.Element;
+export declare const InstalledBadge: ({ item }: {
+    item: ComponentSchema;
+}) => any;
+export declare const ResultCard: ({ result, isSelected, }: {
+    result: {
+        item: ComponentSchema;
+    };
+    isSelected: boolean;
+}) => import("react/jsx-runtime").JSX.Element;
+export declare const CategorySectionBadge: ({ item }: {
+    item: ComponentSchema;
+}) => import("react/jsx-runtime").JSX.Element;
+export declare const ComponentAccessType: ({ item }: {
+    item: ComponentSchema;
+}) => import("react/jsx-runtime").JSX.Element;
+export declare const ResultsCounter: () => import("react/jsx-runtime").JSX.Element;
+export declare const InstallingSpinnerLabel: () => import("react/jsx-runtime").JSX.Element;
+export declare const UsageBanner: () => import("react/jsx-runtime").JSX.Element;
+//# sourceMappingURL=index.d.ts.map
