+++ code/packages/bento-get/types/screens/InstallConfirmScreen.d.ts
@@ -1,0 +1,2 @@
+export declare const InstallConfirmScreen: () => import("react/jsx-runtime").JSX.Element | null;
+//# sourceMappingURL=InstallConfirmScreen.d.ts.map
