+++ code/packages/bento-get/types/screens/SearchScreen.d.ts
@@ -1,0 +1,2 @@
+export declare const SearchScreen: () => import("react/jsx-runtime").JSX.Element;
+//# sourceMappingURL=SearchScreen.d.ts.map
